offline install update

+++ airgap/kubespray-offline/astrago_image_push.sh
@@ -4,18 +4,18 @@ REGISTRY_IMAGE=registry:2.8.2
 REGISTRY_DIR=$(pwd)/outputs/registry-volume
 REGISTRY_PORT=35000
 LOCAL_REGISTRY=localhost:${REGISTRY_PORT}
-# 스크립트 실행 시 에러 발생 시 종료
+# Exit the script if any command fails
 set -e
 
-# containerd와 nerdctl 설치 함수
+# Function to install containerd and nerdctl
 install_containerd_and_nerdctl() {
-    # containerd 설치 여부 확인
+    # Check if containerd is already installed
     if command -v containerd &> /dev/null; then
-        echo "containerd가 이미 설치되어 있습니다. 설치를 스킵합니다."
+        echo "containerd is already installed. Skipping installation."
     else
-        echo "containerd 설치를 시작합니다..."
+        echo "Starting containerd installation..."
 
-        # 필요한 패키지 업데이트 및 설치
+        # Update and install necessary packages
         sudo apt-get update
         # Add Docker's official GPG key:
         sudo apt-get install ca-certificates curl
@@ -28,45 +28,45 @@ install_containerd_and_nerdctl() {
           "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu \
           $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
         sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
-        sudo apt-get update        
+        sudo apt-get update
         sudo apt-get install -y containerd.io
 
-        # 기본 설정 파일 생성
+        # Create default configuration file
         sudo mkdir -p /etc/containerd
         sudo containerd config default | sudo tee /etc/containerd/config.toml
 
-        # containerd 서비스 시작 및 부팅 시 자동 시작 설정
+        # Start containerd service and enable it to start on boot
         sudo systemctl restart containerd
         sudo systemctl enable containerd
 
-        echo "containerd가 성공적으로 설치 및 설정되었습니다."
+        echo "containerd has been successfully installed and configured."
     fi
 
-    # nerdctl 설치 여부 확인
+    # Check if nerdctl is already installed
     if command -v nerdctl &> /dev/null; then
-        echo "nerdctl이 이미 설치되어 있습니다. 설치를 스킵합니다."
+        echo "nerdctl is already installed. Skipping installation."
     else
-        echo "nerdctl 설치를 시작합니다..."
+        echo "Starting nerdctl installation..."
         NERDCTL_VERSION=$(curl -s https://api.github.com/repos/containerd/nerdctl/releases/latest | grep -Po '"tag_name": "\K.*?(?=")')
         curl -LO https://github.com/containerd/nerdctl/releases/download/${NERDCTL_VERSION}/nerdctl-${NERDCTL_VERSION#v}-linux-amd64.tar.gz
         sudo tar Cxzvf /usr/local/bin nerdctl-${NERDCTL_VERSION#v}-linux-amd64.tar.gz
         rm nerdctl-${NERDCTL_VERSION#v}-linux-amd64.tar.gz
 
-        echo "nerdctl가 성공적으로 설치되었습니다."
+        echo "nerdctl has been successfully installed."
     fi
 }
 
-# 프라이빗 Docker 레지스트리 실행 함수
+# Function to run private Docker registry
 run_private_registry() {
-    echo "프라이빗 Docker 레지스트리를 실행합니다..."
+    echo "Running private Docker registry..."
 
-    # Registry가 이미 실행 중인지 확인
+    # Check if Registry is already running
     if nerdctl ps --format '{{.Names}}' | grep -q '^registry$'; then
-        echo "Registry가 이미 실행 중입니다. 스킵합니다."
+        echo "Registry is already running. Skipping."
         return
     fi
 
-    # nerdctl을 사용하여 프라이빗 레지스트리 실행
+    # Run private registry using nerdctl
     sudo nerdctl run -d \
         --name registry \
         -p $REGISTRY_PORT:5000 \
@@ -79,22 +79,51 @@ run_private_registry() {
         -v $REGISTRY_DIR:/var/lib/registry \
         $REGISTRY_IMAGE
 
-    echo "프라이빗 Docker 레지스트리가 성공적으로 실행되었습니다."
-    echo "레지스트리는 localhost에서 접근 가능합니다."
+    echo "Private Docker registry has been successfully started."
+    echo "Registry is accessible on localhost."
+}
+
+# Function to check if image exists in private registry
+check_image_exists() {
+    local image_name="$1"
+    local tag="$2"
+
+    response=$(curl -s -o /dev/null -w "%{http_code}" \
+        -H "Accept: application/vnd.oci.image.manifest.v1+json" \
+        -H "Accept: application/vnd.oci.image.index.v1+json" \
+               "http://${LOCAL_REGISTRY}/v2/${image_name}/manifests/${tag}")
+
+    if [ "$response" = "200" ]; then
+        return 0
+    else
+        return 1
+    fi
 }
 
 pull_and_push_images() {
     images=$(cat outputs/images/*.list)
     for image in $images; do
-
         # Removes specific repo parts from each image for kubespray
         newImage=$image
-	
-        for repo in registry.k8s.io k8s.gcr.io gcr.io docker.io quay.io "nvcr.io/nvidia/cloud-native" "nvcr.io/nvidia/k8s" "nvcr.io/nvidia"; do
+
+        for repo in registry.k8s.io k8s.gcr.io gcr.io docker.io quay.io "nvcr.io/nvidia/cloud-native" "nvcr.io/nvidia/k8s" "nvcr.io/nvidia" "ghcr.io"; do
             newImage=$(echo ${newImage} | sed s@^${repo}/@@)
         done
 
-        newImage=${LOCAL_REGISTRY}/${newImage}	
+        # Separate image name and tag
+        image_name=$(echo ${newImage} | cut -d: -f1)
+        tag=$(echo ${newImage} | cut -d: -f2)
+        if [ "$tag" = "$image_name" ]; then
+            tag="latest"
+        fi
+
+        # Check if image already exists in private registry
+        if check_image_exists "$image_name" "$tag"; then
+            echo "===> Image ${image_name}:${tag} already exists in private registry. Skipping."
+            continue
+        fi
+
+        newImage=${LOCAL_REGISTRY}/${newImage}
 
         echo "===> Pull ${image}"
         sudo nerdctl pull ${image} || exit 1
@@ -103,16 +132,17 @@ pull_and_push_images() {
         sudo nerdctl tag ${image} ${newImage} || exit 1
 
         echo "===> Push ${newImage}"
-        sudo nerdctl push ${newImage} || exit 1	
+        sudo nerdctl push ${newImage} || exit 1
 
-        if [[ $image != $REGISTRY_IMAGE ]]; then
-            echo "===> Remove ${image} and ${newImage}"	  
+        if [[ $image != *$REGISTRY_IMAGE ]]; then
+            echo "===> Remove ${image} and ${newImage}"
             sudo nerdctl rmi ${image} ${newImage}
-	fi
+        fi
     done
 }
 
-# 함수 호출
+# Function calls
+cat imagelists/*.txt | sed "s/#.*$//g" | sort -u > outputs/images/additional-images.list
 install_containerd_and_nerdctl
 run_private_registry
 pull_and_push_images
